fix the order of Mat::at

+++ jsk_pcl_ros/src/depth_image_error_nodelet.cpp
@@ -59,7 +59,7 @@ namespace jsk_pcl_ros
     cv_ptr = cv_bridge::toCvCopy(depth_image, sensor_msgs::image_encodings::TYPE_32FC1);
     cv::Mat cv_depth_image = cv_ptr->image;
     double depth_from_depth_sensor = cv_depth_image.at<float>((int)uv_point->point.y, (int)uv_point->point.x);
-    NODELET_INFO("(u, v) = (%f, %f)", uv_point->point.x, uv_point->point.y);
+    NODELET_INFO("(u, v) = (%d, %d)", (int)uv_point->point.x, (int)uv_point->point.y);
     NODELET_INFO("(z, d) = (%f, %f)", uv_point->point.z, depth_from_depth_sensor);
   }
 }
